Swapped out WebSocket implementation.
Ours does not seems to no longer behave correctly. We have issues with Gonimo
not loading in both the Warp version and the Android version. Ghcjs variant was
not affected.

It seems that the onOpen signal is not triggered reliably or something, because
the client sometimes does not authenticate at all.

Probably some multithreading issue as it does only happen sometimes (very seldom
on Android, quite frequently on Warp) and because it only happens for Haskell
runtimes that run on their own.

We for now swap out our implementation with the one in reflex, as it does not
seem to suffer from this issue. Let's see whether we will have problems again
with:

- opened gonimo in another tab (although it wasn't)
- very long reconnect times