Introduced RIO monad.
Use a concrete monad (Reader + IO) with a polymorphic Reader environment.

This way we have full access to IO when needed. Also logging is now accomplished
by means of a logging function which resides in the reader environment.

Lots of cleanup was possible due to this change, especially Auth code.